Add Path Objects to Mojolicious in lenghtwise order from shortest to longest path.
Currently the Swagger2 Path Objects are added to the Mojolicious routing table
in a random order.
This patch fixes the ordering to be from shortest path to longest:

eg:
/pets/status is added before /pets/{petnumber}

The order of adding routes to the Mojolicious roting table is important, so it
is good to be consistent about it in the Swagger2-plugin as well.