Avoid skipping unresolved $refs
This patch fixes an issue in _resolve_schema that can cause skipping an unresolved
$ref.

JSON::Validator has a functionality to avoid resolving a same reference twice by
storing the memory address of a reference to remember that it has already been
resolved before. However, it seems that at some point that reference seems to be
destroyed and that memory address will be freed for another reference. This causes
a bug where _resolve_schema thinks it has already seen that reference, but in
reality that memory address now holds a different reference.

The issue can be fixed simply by storing the reference.

This issue was noticed with a relatively large OpenAPI specification of Koha ILS
that is split over multiple files. By inserting some debug warnings to print out
the contents of stored reference, we noticed the following:

Current reference in memory address 32881992 is
{
  'description' => 'Bad username/cardnumber and/or password',
  'schema' => {
    '$ref' => '../definitions.json#/error'
  }
}

Last seen in memory address 32881992 was
{
  'type' => [
    'string',
    'null'
  ],
  'description' => 'Accountline note'
}

Since memory address 32881992 was already marked as seen, the "Current reference"
in that memory address will be skipped and the $ref will be left unresolved.

To test:
1. Run t/openapi-unresolved-ref.t