Test problem seen when directory name has RFC 3986 reserved chars
The "definitions" file has a space here, but the problem shows up (thanks
to the call to Mojo::File::realpath) when any character that Mojo::URL
percent-encodes is present anywhere in the full directory path. This is
because the encoded version is then used as if it is a viable filesystem
location, but it is not due to the encoding.