protect against XXE DOS attacks on external refs
there's a new(ish) kind of DOS attack in the OWASP top ten: XXE[1],
where referring to an external entity that never returns can lead to DOS
attacks when parsing XML documents as they try to load those
never-ending external entities.

although they refer to XML there, it's pretty clear this can also affect
JSON-Schema and here's a test case that shows this. the fix has been
made in Validator.pm to prevent loading of files from /dev - this is a
pretty crude fix, which perhaps should be replaced with something more
generic like a timeout?

note there are more interesting attacks if we reference /etc/passwd:

    Can't use string ("nobody:*:-2:-2:Unprivileged User"...) as a
        HASH ref while "strict refs" in use at
        lib/JSON/Validator.pm line 269.

this is probably to discuss, but for now here's a trivial test case +
crude fix that show the POC in action

[1] https://www.owasp.org/index.php/XML_External_Entity_(XXE)_Processing